Finalizando o módulo supplier

+++ src/main/java/com/distribuidora/system_oficina/category/service/CategoryService.java
@@ -1,4 +1,5 @@
 package com.distribuidora.system_oficina.category.service;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import com.distribuidora.system_oficina.category.dto.CategoryRequestDTO;
 import com.distribuidora.system_oficina.category.dto.CategoryResponseDTO;
@@ -7,6 +8,7 @@ import com.distribuidora.system_oficina.category.repository.CategoryRepository;
 import lombok.RequiredArgsConstructor;
 import java.util.List;
 import java.util.stream.Collectors;
+import org.springframework.web.server.ResponseStatusException;
 
 @RequiredArgsConstructor
 @Service
@@ -29,19 +31,22 @@ public class CategoryService {
                 .collect(Collectors.toList());
     }
     public CategoryResponseDTO getCategoryById(Integer id){
-        return toResponseDTO(categoryRepository.findById(id).orElseThrow(() -> new RuntimeException("Category not found")));
+        return toResponseDTO(categoryRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Category not found")));
     }
     public CategoryResponseDTO createCategory(CategoryRequestDTO dto){
         Category category = toEntity(dto);
         return toResponseDTO(categoryRepository.save(category));
     }
     public CategoryResponseDTO updateCategory(Integer id, CategoryRequestDTO dto){
-        Category category = categoryRepository.findById(id).orElseThrow(() -> new RuntimeException("Category not found"));
+        Category category = categoryRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Category not found"));
             category.setName(dto.getName());
             category.setDescription(dto.getDescription());
         return toResponseDTO(categoryRepository.save(category));
     }
     public void deleteCategory(Integer id){
+        if (!categoryRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Category not found");
+        }
         categoryRepository.deleteById(id);
     }
 }
